Set repo directory as safe in the git global config
About a month ago a Git security vulnerability was annouced:
https://github.blog/2022-04-12-git-security-vulnerability-announced/
The security patch to address this vulnerability prevents Git from
looking at .git folders owned by other users, unless these directories
are added to a "safe" list in the Git config.
This Github Action uses a Docker container based on Ubuntu 20.04, and
the version of Git shipped with Ubuntu 20.04 was also patched. As a
result, it is now necessary for this Action to set the repo directory as
safe *inside the container*, given that the owner of the repo directory
(Github runner user) is different from the Docker user
(root). Otherwise, all Git commands fail.

See also https://github.com/actions/checkout/issues/760.

Fixes #9

Signed-off-by: Antonin Bas <abas@vmware.com>